Split configuration for Authenticator and Signer
