+++ mauth-authenticator/src/main/resources/reference.conf
@@ -1,8 +1,3 @@
-app {
-  uuid: ${?APP_UUID}
-  private_key: ${?APP_PRIVATE_KEY}
-}
-
 mauth {
   base_url: ${?MAUTH_URL}
   request_url: "/mauth/v1"
