+++ mauth-proxy/pom.xml
@@ -4,7 +4,7 @@
     <parent>
         <artifactId>mauth-client</artifactId>
         <groupId>com.mdsol.clients</groupId>
-        <version>2016.3.0-SNAPSHOT</version>
+        <version>2016.3.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
     <artifactId>mauth-proxy</artifactId>
