+++ mauth-signer/src/main/resources/reference.conf
@@ -1,0 +1,4 @@
+app {
+  uuid: ${?APP_UUID}
+  private_key: ${?APP_PRIVATE_KEY}
+}
