+++ mauth-test-utils/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <artifactId>mauth-client</artifactId>
         <groupId>com.mdsol.clients</groupId>
-        <version>2016.3.0-SNAPSHOT</version>
+        <version>2016.3.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
