+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>com.mdsol.clients</groupId>
     <artifactId>mauth-client</artifactId>
-    <version>2016.3.0-SNAPSHOT</version>
+    <version>2016.3.1-SNAPSHOT</version>
     <packaging>pom</packaging>
 
     <name>mauth-client</name>
